@@ -1,13 +1,13 @@
 [tool.poetry]
 name = "SimpleLLMFunc"
-version = "0.5.0.beta1"
+version = "0.5.0.beta2"
 description = "A lightweight yet complete LLM/Agent application development framework. Provides decorators that use function docstrings as prompts, requiring no function body implementation while allowing you to benefit from function definitions and type annotations for higher development efficiency. Seamlessly integrate LLM capabilities into any Python project with minimal code."
 authors = ["Ni Jingzhe <nijingzhe@zju.edu.cn>"]
 readme = "README.md"
 packages = [{ include = "SimpleLLMFunc", from = "." }]
 
 [tool.poetry.dependencies]
-python = ">=3.11,<4.0"
+python = ">=3.12,<4.0"
 openai = ">=1.84.0,<2.0.0"
 pyyaml = ">=6.0.2,<7.0.0"
 pydantic = ">=2.11.5,<3.0.0"
@@ -26,6 +26,8 @@ types-tqdm = "^4.67.0.20250809"
 sphinx-intl = "^2.3.2"
 myst-parser = "^4.0.1"
 pytest-asyncio = "^1.3.0"
+hmr = "^0.7.6.2"
+hmr-daemon = "^0.4.2.1"
 
 [build-system]
 requires = ["poetry-core>=2.0.0,<3.0.0"]
